Fix fclCombine task parameter widths

+++ src/Event/Fcl/fclCombine.c
@@ -45,7 +45,7 @@ u8 * FUN_003d2740(u64 param_1,u64 param_2);
 u32 FUN_003d2c10(long param_1);
 #pragma alias FUN_003d2f00_result FUN_003d2f00
 extern s32 FUN_003d2f00_result(void);
-u64 FUN_003d32e0(u64 param_1,u64 param_2);
+u64 FUN_003d32e0(u64 param_1,u32 param_2);
 void FUN_003d3760(int param_1,long param_2,int param_3);
 void FUN_003d38b0(u64 param_1,int param_2,int param_3);
 u32 FUN_003d3b70(long param_1,u32 param_2);
@@ -4833,7 +4833,7 @@ void FUN_003d2f00(void)
 // FUN_003D32E0 NONMATCHING
 
 
-u64 FUN_003d32e0(u64 param_1,u64 param_2)
+u64 FUN_003d32e0(u64 param_1,u32 param_2)
 
 
 
@@ -4847,7 +4847,7 @@ u64 FUN_003d32e0(u64 param_1,u64 param_2)
 
   u32 uVar4;
 
-  u64 uVar5;
+  u32 uVar5;
 
   u64 uVar6;
